moved nmat_mv and nmat_mm tests into ctest_mat. Changed logic in nmat_mm and nmat_mv to throw error when user tries to pass some nmats on gpu and some on cpu
